feat: Add Gist configuration for data synchronization
Introduces a new constant `PUBLIC_GIST_RAW_URL` in `App.tsx` to allow users to specify a raw URL for their public GitHub Gist. This Gist will be used for synchronizing application data.

The `SettingsModal` component's `DataManagementTab` has been updated to accept this `publicGistUrl` as a prop and utilize it for Gist operations, replacing the previous direct reliance on `localStorage.getItem('gistUrl')`. This change facilitates a more robust and user-configurable data backup and synchronization mechanism.